Add supporter finder and tests
We take the first created supporter when multiple records have
the same email. This was decided after talking to the supporter
team.

In Salesforce, there is a dupe calculator that assumes two contacts
are the same if they have two repeated fields. When this happens,
the supporter team merges the duplicates against the first created
one.