Add donation creator and validator
For now we are leaving some fields hardcoded. There are four
fields that are required to create a donation in Salesforce:

 - Amount: comes from the request
 - CloseDate: comes from the gateway, harcoded for now
 - Name: harcoded but is different for recurring donations
 - StageName: hardcoded but is different for recurring donations

The Salesforce API allows to create an opportunity without it
being associated to an account. So these fields alone allow to
create a one-off donation in Salesforce.

But for our application we need to ensure that the donation is
attached to a supporter, so we need the account id from the
supporter. That means that the account id is a required field
for the app.