Add the Stripe gem
We are using the Stripe gem, which requires obtaining a token
in the frontend using AJAX in JS.

The reason for moving to this approach is that before we thought
that we could be able to avoid using JavaScript and do all
the payment handling from the server. However that would
make PCI compliance more strict, so we changed our minds.

In addition, Radar, which is Stripe's integrated solution for
automatic fraud protection, only supports integrations that
make use of client-side tokenization.