Make a gem of the donation system
The gemspec follows the documentation in:

    http://guides.rubygems.org/specification-reference/

The versions of the gems were specified in the gemspec
file to avoid the warnings thrown by the command `gem build`.
I left the restforce gem as it was because that is how the
gem's documentation suggests you to use it.

The Gemfile.lock was gitignored to prevent from accidentally
pushing it in the future.