Prepare for using the request data and payment data
Salesforce needs some of the information coming from the
payment, hence we now have to pass both the request data and
the payment data.

We are now using a request data object like the one sent by
the consumer of the class.

Also we were sending emails in the payment tests so now the mailer
is stubbed.