Add the rest of fields to Salesforce supporters
This adds the optional parameters (i.e., those that are not
required in our Salesforce configuration) to the Supporter
table.

Some fields that are used in the old site were left out:

  * FirstName: we have to first separate both from the name
  field

  * Web_Login__c: we decided that we don't want to save
  supporter data in our database

  * OwnerId: it is defined in maloca as
  SalesforceUser.current_user.Id, but at this stage we don't
  have any way to deal with a Salesforce user.

  * Email_is_Web_Login_legacy__c: the new website won't have a
  login

Also, at the moment LastName is still using the full name until
we find a good way of separating it. I am using placeholders for
the greeting for the same reason. Maloca's code is also
truncating some fields, which for now we are not doing either.