Fix a new rubocop rule introduced by a new version of rubocop
Since this is a gem it is a best practice to not include a Gemfile.lock, however
that also means that travis will download always the last versions of all gems.

In the last Travis run, the version of Rubocop was updated, throwing an error in
a rule that we may not neccessarily want to adopt, hence I am excluding it just
for the file that triggers it.

Relevant Rubocop issue:
  https://github.com/bbatsov/rubocop/issues/3438