Create a fields generator
This is the first step in separating the request data validation
from the Stripe fields generation.

This is done because for recurring donations we have 3 more Stripe
objects (Plan, Customer and Subscription) that need to be passed
different fields. However, the validation is the same for all, since
it deals with the request input data. Hence, this is the right time
to separate the two tasks (validation and field generation), which
now change for different reasons.