Add an input validator with just validation logic
This is the second part of splitting the previous input validator
in two after more Stripe objects having to be created.

Also, since the validation part is the same for all Stripe objects,
there is no point in calling the same object again for each of them,
so instead I moved it one level up and now I can call it once for
all Stripe objects.