Split Salesforce donation validator
This change allows to separate the fields-generation logic from the
validation logic. Now that we have one-off and recurring donations,
the fields will be different, but the input data to be validated is
still the same. So the validation and the fields generation will
change independently, hence this separation.

Some fields that are used in the old site were left out of the
fields generator:

  * Donation_for__c and Web_campaign_reference__c: Both need
  appeals, which are not yet implemented.