Add a payment creator
This is a Paypal-only step.

It creates a payment in Paypal and returns the object.
This payment object can then be used by the consumer of
the gem, for example, to convert it to JSON and send it
to the JavaScript library in the frontend.

Although you can merrily use Payment.new and the gem will
understand you refer to the PayPal gem, I preferred to
rather use PayPal::SDK::REST::DataTypes::Payment to be
more explicit and avoid collisions.