Allow to select other gateways in the payment gateway class
Since the payment selection is configuration, it doesn't make
sense to have configuration mixed with business logic.

To separate the two, a selector is introduced, which
contains the knowledge of the different types of gateways
and their adapters.

Ideally this module could potentially be extended in the
future if we decided to use another supporter database
rather than Salesforce.